requirements.txt
aniso8601==9.0.1
blinker==1.6.2
certifi==2022.12.7
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
importlib-metadata==6.6.0
influxdb-client==1.36.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
reactivex==4.0.4
six==1.16.0
typing_extensions==4.5.0
urllib3==2.0.1
Werkzeug==2.3.3
zipp==3.15.0
